Fix current working dir bug in snaplet initialization
The snap server thread was started correctly in its own sandbox temp
directory.  However, tests were not running in sandbox dir but in
where ever the test suite was started from.  This is a problem because
tests also call into appInit like

    assertGoodUser = do
      let hdl = with auth $ createUser "foo" "foo"
      res <- evalHandler (ST.get "" M.empty) hdl appInit
      either (assertFailure . show) (assertBool failMsg . isRight) res

which means tests get their snaplets initialized in the wrong
directory.