Make tests pass.
The test suite does some invalid reads of non-upgraded
databases.  This exposed the fact that SQLite cannot
decode a NULL-blob into an empty ByteString.
So I have added "blob not null default x''" as
the type for roles and meta.

Some code was also left out from the last commit.